pyproject.toml
[project]
name = "utttrlsim"
version = "0.1.0"
description = "Ultimate Tic-Tac-Toe Reinforcement Learning Simulator"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "gymnasium>=0.29.0",
    "pyyaml>=6.0.2",
    "stable-baselines3>=2.4.1",
    "tensorboard>=2.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "hypothesis>=6.0.0",
]

[project.scripts]
utttrlsim = "utttrlsim.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/utttrlsim"]


[tool.mypy]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=utttrlsim --cov-report=term-missing"

[dependency-groups]
dev = [
    "black>=24.8.0",
    "flake8>=5.0.4",
    "hypothesis>=6.113.0",
    "mypy>=1.14.1",
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
]
